requirements.txt
python-dotenv 
flask
flask-cors
requests
gunicorn
google-cloud-storage